Feature/increase multipart size (#539)
* remove upload limits

* Update Validator.cs

* remove failing tests

+++ api/src/Vote.Monitor.Api/Program.cs
@@ -50,7 +50,7 @@ builder.WebHost.ConfigureKestrel(o =>
 builder.Services.Configure<FormOptions>(options =>
 {
     options.ValueLengthLimit = int.MaxValue;
-    options.MultipartBodyLengthLimit = int.MaxValue;
+    options.MultipartBodyLengthLimit = long.MaxValue;
     options.MultipartHeadersLengthLimit = int.MaxValue;
 });
